Hacked-together support for private modules via replace
The idea is basically just to put
    replace module.name => /path/to/original/module
in the `go.mod`, so that Go will use the copy on-disk instead of
fetching it.  This is a bit more complicated than building in the
current module, but it keeps the tempfiles in the `gomacro.imports`
directory, which is nice.  Again there's some cleanup to do.
See #82 for more context.